requirements.txt
streamlit
trimesh
pandas
numpy
scikit-learn
sentence-transformers
plotly
requests
python-dotenv
pyglet<2